+++ Codigo Fuente/Incidentes/Incidentes.DTOs/TareaDTO.cs
@@ -11,7 +11,7 @@ namespace Incidentes.DTOs
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public int Costo { get; set; }
+        public double Costo { get; set; }
         public int Duracion { get; set; }
         public int ProyectoId { get; set; }
         public TareaDTO() { }
